restructure rules into separate file
